Revert "feat(bot-client): /character alias management — v2 parity phase 1"
This reverts commit b89021e45b3a18a8d8fede19a2021301a3f8cd98.
