+++ docs/commands.md
@@ -14,7 +14,6 @@ Complete reference for every `/` command Tzurot exposes in Discord. For a high-l
 |              | `import` `export` `template`              | Character portability (JSON)                           |
 |              | `chat` `avatar` `voice` `voice-clear`     | Interaction and per-character voice cloning enrollment |
 |              | `settings` `overrides`                    | Per-character config and personal overrides            |
-|              | `alias`                                   | Manage @mention aliases (list, add, remove)            |
 | `/persona`   | `view` `edit` `create` `browse` `default` | User persona management                                |
 |              | `override set` `override clear`           | Per-character persona overrides                        |
